Authorization checks was added to Users controller

+++ public_html/protected/controllers/UsersController.php
@@ -2,48 +2,7 @@
 
 class UsersController extends Controller
 {
-    /**
-     * @var string the default layout for the views. Defaults to '//layouts/column2', meaning
-     * using two-column layout. See 'protected/views/layouts/column2.php'.
-     */
     public $layout = '//layouts/yiistrap.col2';
-
-    /**
-     * @return array action filters
-     */
-//    public function filters()
-//    {
-//        return array(
-//            'accessControl', // perform access control for CRUD operations
-//            'postOnly + delete', // we only allow deletion via POST request
-//        );
-//    }
-
-    /**
-     * Specifies the access control rules.
-     * This method is used by the 'accessControl' filter.
-     * @return array access control rules
-     */
-//    public function accessRules()
-//    {
-//        return array(
-//            array('allow', // allow all users to perform 'index' and 'view' actions
-//                'actions' => array('index', 'view'),
-//                'users' => array('*'),
-//            ),
-//            array('allow', // allow authenticated user to perform 'create' and 'update' actions
-//                'actions' => array('create', 'update'),
-//                'users' => array('@'),
-//            ),
-//            array('allow', // allow admin user to perform 'admin' and 'delete' actions
-//                'actions' => array('admin', 'delete'),
-//                'users' => array('admin'),
-//            ),
-//            array('deny', // deny all users
-//                'users' => array('*'),
-//            ),
-//        );
-//    }
 
     /**
      * Displays a particular model.
@@ -51,9 +10,14 @@ class UsersController extends Controller
      */
     public function actionView($id)
     {
-        $this->render('view', array(
-            'model' => $this->loadModel($id),
-        ));
+        if (Yii::app()->user->checkAccess('admin')) {
+            $this->render('view', array(
+                'model' => $this->loadModel($id),
+            ));
+        }
+        else {
+            throw new CHttpException(403);
+        }
     }
 
     /**
@@ -62,6 +26,10 @@ class UsersController extends Controller
      */
     public function actionCreate()
     {
+        if (!Yii::app()->user->checkAccess('admin')) {
+            throw new CHttpException(403);
+        }
+
         $model = new Users;
 
 // Uncomment the following line if AJAX validation is needed
@@ -85,6 +53,10 @@ class UsersController extends Controller
      */
     public function actionUpdate($id)
     {
+        if (!Yii::app()->user->checkAccess('admin')) {
+            throw new CHttpException(403);
+        }
+
         $model = $this->loadModel($id);
 
 // Uncomment the following line if AJAX validation is needed
@@ -108,6 +80,10 @@ class UsersController extends Controller
      */
     public function actionDelete($id)
     {
+        if (!Yii::app()->user->checkAccess('admin')) {
+            throw new CHttpException(403);
+        }
+
         if (Yii::app()->request->isPostRequest) {
 // we only allow deletion via POST request
             $this->loadModel($id)->delete();
@@ -124,6 +100,10 @@ class UsersController extends Controller
      */
     public function actionIndex()
     {
+        if (!Yii::app()->user->checkAccess('admin')) {
+            throw new CHttpException(403);
+        }
+
         $dataProvider = new CActiveDataProvider('Users');
         $this->render('index', array(
             'dataProvider' => $dataProvider,
@@ -135,6 +115,10 @@ class UsersController extends Controller
      */
     public function actionAdmin()
     {
+        if (!Yii::app()->user->checkAccess('admin')) {
+            throw new CHttpException(403);
+        }
+
         $model = new Users('search');
         $model->unsetAttributes(); // clear any default values
         if (isset($_GET['Users']))
